refactor : improve code quality
