+++ src/main/java/com/thee5176/ledger_command/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.thee5176.ledger_command.config;
+package com.thee5176.ledger_command.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,11 +13,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import com.thee5176.ledger_command.security.CustomUserDetailsService;
-import com.thee5176.ledger_command.security.JOOQAuthoritiesRepository;
-import com.thee5176.ledger_command.security.JOOQUsersRepository;
-import com.thee5176.ledger_command.security.JwtAuthenticationFilter;
-import com.thee5176.ledger_command.security.JwtService;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
@@ -55,7 +50,7 @@ public class WebSecurityConfig {
 		http
 			.csrf((csrf -> csrf.disable()))
 			.authorizeHttpRequests(auth -> 
-				auth.requestMatchers("/api/v1/auth/**", "/error").permitAll()
+				auth.requestMatchers("/api/v1/auth/login","/api/v1/auth/register", "/error").permitAll()
 				.anyRequest().authenticated()
 				)
 			.sessionManagement(session -> session
