+++ src/test/java/com/thee5176/ledger_command/record/application/controller/HealthCheckControllerTest.java
@@ -2,7 +2,6 @@ package com.thee5176.ledger_command.record.application.controller;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.ResponseEntity;
-import com.thee5176.ledger_command.record.application.controller.HealthCheckController;
 
 class HealthCheckControllerTest {
